Split partFuncs into separate functions
