+++ functions/part_funcs/create_all_plots.py
@@ -1,11 +1,5 @@
-import openpyxl
 import numpy as np
-from pptx.util import Inches
-from pptx import Presentation
 import matplotlib.pyplot as plt
-from openpyxl.styles import Alignment
-from openpyxl.styles import PatternFill
-from openpyxl.styles.borders import Border, Side
 
 from functions.plotFuncs import plotPie, plotBar, plotLine, plotStack
 
@@ -365,148 +359,3 @@ def create_all_plots(myWorkbook, part_label, results_dir, start_label,
     plt.close("all")
 
     return spendings_list, incomes_list
-
-
-def create_pptx_presentation(part_label, results_dir):
-
-    prs = Presentation()
-    title_slide_layout = prs.slide_layouts[0]
-    blank_slide_layout = prs.slide_layouts[6]
-    slides = list()
-
-    slides.append(prs.slides.add_slide(title_slide_layout))
-    title = slides[-1].shapes.title
-    title.text = part_label + " - raport finansowy"
-
-    # Part as a whole
-    slides.append(prs.slides.add_slide(title_slide_layout))
-    title = slides[-1].placeholders[1]
-    title.text = "1. Cały przedział"
-
-    for i in range(6):
-        slides.append(prs.slides.add_slide(blank_slide_layout))
-        left = Inches(1.5)
-        top = Inches(0.0)
-        height = width = Inches(7.5)
-
-        pic_path = results_dir + "/plots/plot" + str(i + 1) + ".png"
-        pic = slides[-1].shapes.add_picture(pic_path, left, top, height, width)
-
-    # Averaged month
-    slides.append(prs.slides.add_slide(title_slide_layout))
-    title = slides[-1].placeholders[1]
-    title.text = "2. Uśredniony miesiąc"
-
-    for i in range(4):
-        slides.append(prs.slides.add_slide(blank_slide_layout))
-        left = Inches(1.5)
-        top = Inches(0.0)
-        height = width = Inches(7.5)
-
-        pic_path = results_dir + "/plots/plot" + str(i + 1 + 6) + ".png"
-        pic = slides[-1].shapes.add_picture(pic_path, left, top, height, width)
-
-    # Part as a sequence of months
-    slides.append(prs.slides.add_slide(title_slide_layout))
-    title = slides[-1].placeholders[1]
-    title.text = "3. Sekwencja miesięcy"
-
-    for i in range(5):
-        slides.append(prs.slides.add_slide(blank_slide_layout))
-        left = Inches(0.0)
-        top = Inches(0.1)
-        height = Inches(10.5)
-        width = Inches(7.0)
-
-        pic_path = results_dir + "/plots/plot" + str(i + 1 + 10) + ".png"
-        pic = slides[-1].shapes.add_picture(pic_path, left, top, height, width)
-
-    prs.save(results_dir + "/" + part_label + " - raport finansowy.pptx")
-
-
-def create_xlsx_report(results_dir, part_label, myWorkbook, start_label,
-                       n_of_months):
-
-    wb_to_export = openpyxl.Workbook()
-
-    def export_to_excel(cat_name, num_of_ws):
-        values = myWorkbook.spends_values_yr[cat_name]
-        items = myWorkbook.spends_items_yr[cat_name]
-        monthlabels = myWorkbook.spends_monthlabel_yr[cat_name]
-
-        mdict = {1: "Styczeń", 2: "Luty", 3: "Marzec", 4: "Kwiecień", 5: "Maj",
-                 6: "Czerwiec", 7: "Lipiec", 8: "Sierpień", 9: "Wrzesień",
-                 10: "Październik", 11: "Listopad", 12: "Grudzień"}
-
-        values_dict = dict()
-        items_dict = dict()
-
-        for month_num in np.unique(monthlabels):
-            _one_month_list = [values[i] for i, j in enumerate(monthlabels)
-                               if j == month_num]
-            values_dict[month_num] = _one_month_list
-
-            _one_month_list = [items[i] for i, j in enumerate(monthlabels)
-                               if j == month_num]
-            items_dict[month_num] = _one_month_list
-
-        ws = wb_to_export.create_sheet(cat_name, num_of_ws)
-
-        # "try" becase doesn"t work when no spendings
-        try:
-            ws.column_dimensions["A"].width = max([len(i) for i in items]) + 2
-        except:
-            pass
-
-        row = 0
-        month = start_label[0]
-        year = start_label[1]
-        for number in range(1, n_of_months + 1):  # list(values_dict.keys()):
-            if number in monthlabels:
-                row += 1
-                ws.cell(row, 1).value = mdict[month] + "  " + str(month) \
-                                        + ".20" + str(year)
-                ws.merge_cells(start_row=row, start_column=1, end_row=row,
-                               end_column=2)
-                ws.cell(row, 1).alignment = Alignment(horizontal="center")
-                ws.cell(row, 1).fill = PatternFill(fgColor="93e1e6",
-                                                   fill_type="solid")
-
-                row += 1
-                ws.cell(row, 1).value = "Co?"
-                ws.cell(row, 2).value = "Kwota [zł]"
-                ws.cell(row, 1).alignment = Alignment(horizontal="center")
-                ws.cell(row, 2).alignment = Alignment(horizontal="center")
-                ws.cell(row, 1).fill = PatternFill(fgColor="daf6f7",
-                                                   fill_type="solid")
-                ws.cell(row, 2).fill = PatternFill(fgColor="daf6f7",
-                                                   fill_type="solid")
-
-                for i, val in enumerate(values_dict[number]):
-                    row += 1
-                    ws.cell(row, 1).value = items_dict[number][i]
-                    ws.cell(row, 2).value = val
-
-            month += 1
-            if month > 12:
-                month -= 12
-                year += 1
-
-        thin_border = Border(left=Side(style="thin"),
-                             right=Side(style="thin"),
-                             top=Side(style="thin"),
-                             bottom=Side(style="thin"))
-        for r in range(row):
-            ws.cell(row=r + 1, column=1).border = thin_border
-            ws.cell(row=r + 1, column=2).border = thin_border
-
-        return wb_to_export
-
-    wb_to_export.remove_sheet(wb_to_export.active)
-    wb_to_export = export_to_excel("Rzeczy i sprzęty", 0)
-    wb_to_export = export_to_excel("Hobby i przyjemności", 1)
-    wb_to_export = export_to_excel("Transport i noclegi", 2)
-    wb_to_export = export_to_excel("Podróże", 3)
-
-    wb_to_export.save(results_dir + "/" + part_label
-                      + " - zestawienie wydatków.xlsx")
